kirana-backend-python/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
marshmallow==3.20.1
python-dotenv==1.0.0
pymongo==4.6.0
